streamlit/requirements.txt
numpy==1.19.0
Keras==2.4.2
Pillow==7.1.2
